thermo_data_generator/pyproject.toml
[tool.poetry]
name = "thermo-data-generator"
version = "0.1.0"
description = ""
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
